Crate trie_map

Modules

base16

This module contains the TrieMapBase16 and TrieSetBase16 types, which are specialized versions of TrieMap and TrieSet that use base-16 encoding for their keys.

iterators

This module contains the iterators for TrieMap and TrieSet.

Structs

TrieMap

A trie-based Map keyed on sequences of bytes with values stored at the leaves of the trie. RANGE controls the size of the child array in each node. And BASE_CHAR is the byte value of the first character in the range of characters that can be used in keys. It is used to adjust each character in a key to serve as indices into the their descendant arrays.

TrieSet

A set implemented using TrieMap. Internally this is a TrieMap with a unit type as the value. This is a convenience type for when you want to use a TrieMap as a set.