trie_map

Module base16

This module contains the TrieMapBase16 and TrieSetBase16 types, which are specialized versions of TrieMap and TrieSet that use base-16 encoding for their keys.

Modules

iterators

This module contains the iterators for TrieMapBase16 and TrieSetBase16.

Structs

TrieMapBase16

A trie whose keys can be any string with no restrictions on characters. This is achieved by base 16 encoding the key as it’s inserted into the trie. The encoding is done via an iterator that only encodes the number of bytes needed to determine if a key is missing or present. So determining a miss is as fast as possible for operations such as contains_key().

TrieSetBase16

A set implemented as a trie that encodes its keys as base 16 bytes.