trie_map::base16

Struct TrieMapBase16

pub struct TrieMapBase16<V> { /* private fields */ }

A trie whose keys can be any string with no restrictions on characters. This is achieved by base 16 encoding the key as it’s inserted into the trie. The encoding is done via an iterator that only encodes the number of bytes needed to determine if a key is missing or present. So determining a miss is as fast as possible for operations such as contains_key().

Each node has a fixed-size array for 16 successors. And a value of any type can be inserted with a key. The values are stored in independent contiguous memory, while the terminal trie nodes hold handles to their values.

Implementations

impl<V> TrieMapBase16<V>

pub fn new() -> Self

Creates an empty TrieMapBase16.

pub fn clear(&mut self)

Removes all key-value pairs from the map.

pub fn contains_key(&self, key: &str) -> bool

Returns true if the map contains a value for the specified key.

pub fn contains_key_by_iter<K>(&self, key: K) -> bool
where K: Iterator<Item = u8>,

Returns true if the map contains a value for the specified key.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("Γειά σου", 1);
trie.insert("κόσμος", 2);

assert_eq!(trie.contains_key_by_iter("Γειά σου".bytes()), true);
assert_eq!(trie.contains_key_by_iter("κόσμος".bytes()), true);

pub fn get(&self, key: &str) -> Option<&V>

Returns a reference to the value corresponding to the key, or None if the key is not present in the map.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("Γειά σου", 1);
trie.insert("κόσμος", 2);
trie.insert("👋", 42);
trie.insert("🌍", 43);

assert_eq!(trie.get("Γειά σου"), Some(&1));
assert_eq!(trie.get("κόσμος"), Some(&2));
assert_eq!(trie.get("👋"), Some(&42));
assert_eq!(trie.get("🦀"), None);

pub fn get_by_iter<K>(&self, key: K) -> Option<&V>
where K: Iterator<Item = u8>,

Returns a reference to the value corresponding to the key, or None if the key is not present in the map.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("Γειά σου", 1);

assert_eq!(trie.get_by_iter("Γειά σου".bytes()), Some(&1));

pub fn get_mut(&mut self, key: &str) -> Option<&mut V>

Returns a mutable reference to the value corresponding to the key, or None if the key is not present in the map.

pub fn get_mut_by_iter<K>(&mut self, key: K) -> Option<&mut V>
where K: Iterator<Item = u8>,

Returns a mutable reference to the value corresponding to the key, or None if the key is not present in the map.

pub fn get_or_insert<K>(&mut self, key: K, value: V) -> &mut V
where K: Borrow<str>,

If the key-value pair is not present in the map, inserts it and returns a mutable reference to the value. If the key-value pair is present, returns a mutable reference to the already present value.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("Γειά σου", 1);
trie.insert("κόσμος", 2);
trie.insert("👋", 42);

assert_eq!(trie.get_or_insert("Γειά σου", 0), &mut 1);
assert_eq!(trie.get_or_insert("κόσμος", 0), &mut 2);
assert_eq!(trie.get_or_insert("👋", 0), &mut 42);
assert_eq!(trie.get_or_insert("🌍", 0), &mut 0);

pub fn get_or_insert_by_iter<K>(&mut self, key: K, value: V) -> &mut V
where K: Iterator<Item = u8>,

If the key-value pair is not present in the map, inserts it and returns a mutable reference to the value. If the key-value pair is present, returns a mutable reference to the already present value.

pub fn get_or_insert_with<K, F>(&mut self, key: K, f: F) -> &mut V
where K: Borrow<str>, F: FnOnce() -> V,

If the key-value pair is not present in the map, inserts it and returns a mutable reference to the value. If the key-value pair is present, returns a mutable reference to the already present value.

pub fn get_or_insert_by_iter_with<K, F>(&mut self, key: K, f: F) -> &mut V
where K: Iterator<Item = u8>, F: FnOnce() -> V,

If the key-value pair is not present in the map, inserts it and returns a mutable reference to the value. If the key-value pair is present, returns a mutable reference to the already present value.

pub fn insert<K>(&mut self, key: K, value: V) -> Option<V>
where K: Borrow<str>,

Inserts a key-value pair into the map. If the key already had a value present in the map, that value is returned. Otherwise None is returned.

pub fn iter(&self) -> Iter<'_, V>

Returns an iterator over the key-value pairs of the map in canonical order.

pub fn iter_mut(&mut self) -> IterMut<'_, V>

Returns an iterator over the key-value pairs of the map. The values are mutable.

pub fn keys(&self) -> Keys<'_, V>

Returns an iterator over the keys of the map in canonical order.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("hello", 1);
trie.insert("world", 2);

let mut iter = trie.keys().rev();
assert_eq!(iter.next(), Some("world".to_string()));

pub fn len(&self) -> usize

Returns the number of key-value pairs in the map.

pub fn is_empty(&self) -> bool

Returns true if the map contains no key-value pairs.

pub fn remove(&mut self, key: &str) -> Option<V>

Removes a key from the map, returning the value at the key if the key was previously in the map. If the key was not present in the map, None is returned.

use trie_map::base16::TrieMapBase16;

let mut trie = TrieMapBase16::new();

trie.insert("hello", 1);
trie.insert("world", 2);

assert_eq!(trie.remove("hello"), Some(1));
assert_eq!(trie.remove("hello"), None);

pub fn remove_by_iter<K>(&mut self, key: K) -> Option<V>
where K: Iterator<Item = u8>,

Removes a key from the map, returning the value at the key if the key was previously in the map. If the key was not present in the map, None is returned.

pub fn values(&self) -> Values<'_, V>

Returns an iterator over the values of the map. They will be in the same order as the keys.

pub fn values_mut(&mut self) -> ValuesMut<'_, V>

Returns an iterator over the values of the map. They will be in the same order as the keys.