trie_map::base16

Struct TrieSetBase16

pub struct TrieSetBase16 { /* private fields */ }

A set implemented as a trie that encodes its keys as base 16 bytes.

Implementations

impl TrieSetBase16

pub fn new() -> Self

Creates a new empty TrieSet.

pub fn clear(&mut self)

Removes all elements from the set.

pub fn contains(&self, key: &str) -> bool

Reports whether the set contains the given key.

use trie_map::base16::TrieSetBase16;

let mut set = TrieSetBase16::new();

assert!(set.insert("hello"));

assert!(!set.contains("world"));
assert!(set.contains("hello"));

pub fn contains_by_iter<K>(&self, iter: K) -> bool
where K: Iterator<Item = u8>,

Reports whether the set contains the given key. The key is given as an iterator.

pub fn insert<K>(&mut self, key: K) -> bool
where K: Borrow<str>,

Inserts the given key into the set. Returns true if the key was not already present.

pub fn insert_by_iter<K>(&mut self, iter: K) -> bool
where K: Iterator<Item = u8>,

Inserts the given key into the set. Returns true if the key was not already present.

pub fn is_empty(&self) -> bool

Reports whether the set is empty.

pub fn iter(&self) -> Iter<'_>

Returns an iterator over the set.

pub fn len(&self) -> usize

Returns an iterator over the set.

pub fn remove(&mut self, key: &str) -> bool

Removes the given key from the set. Returns true if the key was present.

pub fn remove_by_iter<K>(&mut self, iter: K) -> bool
where K: Iterator<Item = u8>,

Removes the given key from the set. Returns true if the key was present. The key is given as an iterator.