Module trie_map::iterators

Iterators, The Iterators Used in TrieMap

Each iterator implements the InnerIter trait, which provides the traversal algorithms. The iterators perform a traversal of the trie, advancing through its nodes on each call to next() or next_back(). No intermediate representation is used to store keys or values.

Structs

IntoIter

A consuming iterator over the key-value pairs of a TrieMap.

Iter

An iterator over the key-value pairs of a TrieMap that holds an immutable reference to the trie.

IterMut

An iterator over the key-value pairs of a TrieMap that holds a mutable reference to the trie. It produces items that can be modified by the caller.

Keys

An iterator over the keys of a TrieMap that holds an immutable reference to the trie.

Values

An iterator over the values of a TrieMap that holds an immutable reference to the trie.