trie_map

Struct TrieSet

pub struct TrieSet<const R: usize, const B: u8> { /* private fields */ }

A set implemented using TrieMap. Internally this is a TrieMap with a unit type as the value. This is a convenience type for when you want to use a TrieMap as a set.

Implementations

impl<const R: usize, const B: u8> TrieSet<R, B>

pub fn new() -> Self

Creates a new empty TrieSet.

pub fn clear(&mut self)

Removes all elements from the set.

pub fn contains<K>(&self, key: K) -> bool
where K: AsRef<[u8]>,

Reports whether the set contains the given key.

use trie_map::TrieSet;

let mut set = TrieSet::<26, b'a'>::new();

assert!(set.insert("hello"));
assert!(set.contains("hello"));
assert!(!set.contains("world"));

pub fn contains_by_iter<K>(&self, key: K) -> bool
where K: Iterator<Item = u8>,

Reports whether the set contains the given key. The key is given as an iterator.

use trie_map::TrieSet;

let mut set = TrieSet::<26, b'a'>::new();

assert!(set.insert("hello"));
assert!(set.contains_by_iter("hello".bytes()));
assert!(!set.contains_by_iter("world".bytes()));

pub fn insert<K>(&mut self, key: K) -> bool
where K: AsRef<[u8]>,

Inserts the given key into the set. Returns true if the key was not already present.

use trie_map::TrieSet;

let mut set = TrieSet::<26, b'a'>::new();

assert!(set.insert("hello"));
assert!(!set.insert("hello"));
assert!(set.insert("world"));

pub fn insert_by_iter<K>(&mut self, key: K) -> bool
where K: Iterator<Item = u8>,

Inserts the given key into the set. Returns true if the key was not already present. The key is given as an iterator.

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

pub fn iter(&self) -> Iter<'_, R, B>

Returns an iterator over the keys of the set.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn remove<K>(&mut self, key: K) -> bool
where K: AsRef<[u8]>,

Removes the given key from the set. Returns true if the key was present.

pub fn remove_by_iter<K>(&mut self, key: K) -> bool
where K: Iterator<Item = u8>,

Removes the given key from the set. Returns true if the key was present. The key is given as an iterator.

use trie_map::TrieSet;

let mut set = TrieSet::<26, b'a'>::new();

assert!(set.insert("hello"));
assert!(set.remove_by_iter("hello".bytes()));
assert!(!set.remove_by_iter("hello".bytes()));