Struct trie_map::TrieMap

pub struct TrieMap<V, const RANGE: usize, const BASE_CHAR: u8> { /* private fields */ }

A trie-based Map keyed on sequences of bytes with values stored at the leaves of the trie. RANGE controls the size of the child array in each node. And BASE_CHAR is the byte value of the first character in the range of characters that can be used in keys. It is used to adjust each character in a key to serve as indices into the their descendant arrays.

Implementations

impl<V, const RANGE: usize, const BASE_CHAR: u8> TrieMap<V, RANGE, BASE_CHAR>

pub fn new() -> Self

Creates a new empty trie.

pub fn with_capacity(capacity: usize) -> Self

Creates a new empty trie with the given capacity. Only the internal vector that holds the values is affected by this method.

pub fn clear(&mut self)

Clears the trie, removing all values.

pub fn contains_key<K>(&self, key: K) -> boolwhere K: AsRef<[u8]>,

Returns true if the trie contains a value at the given key, otherwise false is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

assert_eq!(trie.contains_key("hello"), true);

pub fn contains_by_iter<K>(&self, key: K) -> boolwhere K: Iterator<Item = u8>,

Returns true if the trie contains a value at the given key, otherwise false is returned. The trait bound on K is makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter.

use trie_map::TrieMap;
use std::collections::VecDeque;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert(b"hello", 1);

let mut window = b"hhell".iter().copied().collect::<VecDeque<_>>();

window.pop_front();
window.push_back(b'o');

assert_eq!(trie.contains_by_iter(window.range(0..5).copied()), true);

pub fn get<K>(&self, key: K) -> Option<&V>where K: AsRef<[u8]>,

Accesses a value in the trie at the given key, if it exists, a reference is returned, otherwise None is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

assert_eq!(trie.get("hello"), Some(&1));
assert_eq!(trie.get("world"), None);

pub fn get_by_iter<K>(&self, key: K) -> Option<&V>where K: Iterator<Item = u8>,

Accesses a value in the trie at the given key, if it exists, a reference is returned, otherwise None is returned. The trait bound on K is makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter.

use trie_map::TrieMap;
use std::collections::VecDeque;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

let mut window = b"hhell".iter().copied().collect::<VecDeque<_>>();

window.pop_front();
window.push_back(b'o');

assert_eq!(trie.get_by_iter(window.range(0..5).copied()), Some(&1));

pub fn get_mut<K>(&mut self, key: K) -> Option<&mut V>where K: AsRef<[u8]>,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise None is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

*trie.get_mut("hello").unwrap() = 17;

assert_eq!(trie.get("hello"), Some(&17));

pub fn get_mut_by_iter<K>(&mut self, key: K) -> Option<&mut V>where K: Iterator<Item = u8>,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise None is returned. The trait bound on K is makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert(b"hello", 1);

let iter = b"hel".iter().chain(b"lo").copied();

*trie.get_mut_by_iter(iter).unwrap() = 17;

assert_eq!(trie.get(b"hello"), Some(&17));

pub fn get_or_insert<K>(&mut self, key: K, value: V) -> &mut Vwhere K: AsRef<[u8]>,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise the given value is inserted and a reference to it is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

assert_eq!(trie.get_or_insert("hello", 1), &1);

pub fn get_or_insert_by_iter<K>(&mut self, key: K, value: V) -> &mut Vwhere K: Iterator<Item = u8>,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise the given value is inserted and a reference to it is returned. The trait bound on K is makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

let iter = b"hello".iter().chain(b"world").copied();

*trie.get_or_insert_by_iter(iter, 1) = 17;

assert_eq!(trie.get(b"helloworld"), Some(&17));

pub fn get_or_insert_with<K, F>(&mut self, key: K, f: F) -> &mut Vwhere K: AsRef<[u8]>, F: FnOnce() -> V,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise the given function is called to produce a value which is inserted and a reference to it is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

assert_eq!(trie.get_or_insert_with("hello", ||1), &1);

pub fn get_or_insert_by_iter_with<K, F>(&mut self, key: K, f: F) -> &mut Vwhere K: Iterator<Item = u8>, F: FnOnce() -> V,

Returns a mutable reference to a value in the trie at the given key, if it exists, otherwise the given function is called to produce a value which is inserted and a reference to it is returned. The trait bound on K enables passing an iterator or other type that can be converted into an iterator as a paramter.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

let iter = b"hello".iter().chain(b"world").copied();

*trie.get_or_insert_by_iter_with(iter, ||1) = 17;

assert_eq!(trie.get(b"helloworld"), Some(&17));

pub fn insert<K>(&mut self, key: K, value: V) -> Option<V>where K: AsRef<[u8]>,

Inserts a value into the trie at the given key. If the key already exists in the trie, the value is replaced and the old value is returned, otherwise None is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

assert_eq!(trie.get("hello"), Some(&1));

pub fn insert_by_iter<K>(&mut self, key: K, value: V) -> Option<V>where K: Iterator<Item = u8>,

Inserts a value into the trie at the given key. The trait bound on K makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter. If the key already exists in the trie, the value is replaced and the old value is returned, otherwise None is returned.

pub fn is_empty(&self) -> bool

Returns true if the trie contains no values, otherwise false is returned.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::with_capacity(10);

assert_eq!(trie.is_empty(), true);

trie.insert("hello", 1);

assert_eq!(trie.is_empty(), false);

pub fn iter(&self) -> Iter<'_, V, RANGE, BASE_CHAR>

Returns an iterator over the key-value pairs in the trie.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);
trie.insert("world", 2);

let mut iter  = trie.iter();
let     first = iter.next().unwrap();

assert_eq!(first.0.as_ref(), b"hello");
assert_eq!(first.1, &1);

pub fn iter_mut(&mut self) -> IterMut<'_, V, RANGE, BASE_CHAR>

Returns an iterator over the mutable key-value pairs in the trie. Only the values are mutable, not the keys.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);
trie.insert("world", 2);

let mut iter  = trie.iter_mut();
let     first = iter.next().unwrap();
*first.1 = 17;

assert_eq!(trie.get("hello"), Some(&17));

pub fn keys(&self) -> Keys<'_, V, RANGE, BASE_CHAR>

Returns an iterator over the keys in the trie.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert("hello", 1);

assert_eq!(trie.keys().next().unwrap().as_ref(), b"hello");

pub fn len(&self) -> usize

Returns the number of values in the trie.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::with_capacity(10);

trie.insert("hello", 1);

assert_eq!(trie.len(), 1);

pub fn remove<K>(&mut self, key: K) -> Option<V>where K: AsRef<[u8]>,

Removes a value from the trie at the given key, if it exists, and returns it, otherwise None is returned.

pub fn remove_by_iter<K>(&mut self, key: K) -> Option<V>where K: Iterator<Item = u8>,

Removes a value from the trie at the given key, if it exists, and returns it, otherwise None is returned. The trait bound on K makes it easy to pass an iterator or other type that can be converted into an iterator as a paramter as the key.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::with_capacity(10);

trie.insert("hello", 1);

assert_eq!(trie.remove_by_iter("hello".bytes()), Some(1));

pub fn values(&self) -> Values<'_, V, RANGE, BASE_CHAR>

Returns an iterator over the values in the trie.

use trie_map::TrieMap;

let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();

trie.insert(b"hello", 1);

assert_eq!(trie.values().next(), Some(&1));

pub fn values_mut(&mut self) -> ValuesMut<'_, V, RANGE, BASE_CHAR>

Returns an iterator over the values in the trie.

use trie_map::TrieMap;
 
let mut trie: TrieMap<i32, 26, b'a'> = TrieMap::new();
 
trie.insert(b"hello", 1);
 
assert_eq!(trie.values_mut().next(), Some(&mut 1));